* fixinc/inclhack.tpl: Only install assert.h conditionally.         * fixinc/inclhack.sh: Regenerated.         * fixinc/fixincl.sh: Regenerated.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@28659 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/fixinc/fixincl.sh
@@ -406,8 +406,10 @@ done
 #
 # # # # # # # # # # # # # # # # # # # # #
 
-cd $ORIGDIR
-rm -f include/assert.h
-cp ${srcdir}/assert.h include/assert.h || exit 1
-chmod a+r include/assert.h
-
+if [ x${INSTALL_ASSERT_H} != x ] && [ -f ${srcdir}/assert.h ]
+then
+  cd $ORIGDIR
+  rm -f include/assert.h
+  cp ${srcdir}/assert.h include/assert.h || exit 1
+  chmod a+r include/assert.h
+fi
